Working on the getting address from mocking package
