+++ packages/e2e/address.js
@@ -1,0 +1,14 @@
+// @TODO: import this from the mocking package
+
+// Robin: I was struggling making this work
+// const { address1055 } = require("vergunningcheck-mocking/src/address");
+
+const address = {
+  zipCode: "1055XD",
+  streetName: "Louise de Colignystraat",
+  houseNumber: 19,
+  houseNumberAddition: "C",
+  houseNumberFull: "19c",
+};
+
+module.exports = address;
