+++ packages/e2e/nightwatch/config.js
@@ -1,7 +1,9 @@
-const selectors = require("../selectors");
+const address = require("../address");
 const host = require("../host");
+const selectors = require("../selectors");
 
 const config = {
+  address,
   domain: host,
   selectors,
 };
