+++ packages/mocking/src/address.ts
@@ -1,0 +1,7 @@
+export const address1055 = {
+  houseNumber: 19,
+  houseNumberAddition: "C",
+  houseNumberFull: "19c",
+  streetName: "Louise de Colignystraat",
+  zipCode: "1055XD",
+};
