+++ packages/mocking/src/index.ts
@@ -1,2 +1,3 @@
 export * as restriction from "./restriction";
 export * as zipcode from "./zipcode";
+export * as address from "./address";
